# PaperPlane-Media/contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {"title": "PaperPlane Media API", "version": "1.0.0"},
  "servers": [{"url": "https://media.paperplane.codes"}],
  "paths": {
...
    "/v1/admin/tracks": {
      "get": {"operationId": "listAdminTracks", "responses": {"200": {"description": "All tracks", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AdminTrackList"}}}}}},
      "post": {"operationId": "uploadTrack", "requestBody": {"required": true, "content": {"multipart/form-data": {"schema": {"type": "object", "required": ["archive", "title", "artist"], "properties": {"archive": {"type": "string", "format": "binary"}, "title": {"type": "string"}, "artist": {"type": "string"}, "sortOrder": {"type": "integer"}, "enabled": {"type": "boolean"}}}}}}, "responses": {"201": {"description": "Track published", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AdminTrackResponse"}}}}, "413": {"description": "Archive too large"}, "429": {"$ref": "#/components/responses/TooManyRequests"}}},
      "delete": {"operationId": "deleteTracks", "requestBody": {"required": true, "content": {"application/json": {"schema": {"type": "object", "required": ["ids"], "properties": {"ids": {"type": "array", "items": {"type": "integer", "format": "int64"}}}}}}}, "responses": {"200": {"description": "Tracks deleted"}}}
    },
...
  },
  "components": {"responses": {
    "TooManyRequests": {"description": "Rate limit exceeded", "content": {"application/problem+json": {"schema": {"$ref": "#/components/schemas/Problem"}}}}
  }, "schemas": {
...
    "AdminTrack": {"type": "object", "required": ["musicKey", "assetId", "title", "artist", "audioFormat", "sortOrder", "enabled", "createdAt", "updatedAt"], "properties": {"musicKey": {"type": "integer", "format": "int64"}, "assetId": {"type": "string", "format": "uuid"}, "title": {"type": "string"}, "artist": {"type": "string"}, "audioFormat": {"$ref": "#/components/schemas/AudioFormat"}, "coverPath": {"type": ["string", "null"]}, "lyricPath": {"type": ["string", "null"]}, "sortOrder": {"type": "integer"}, "enabled": {"type": "boolean"}, "createdAt": {"type": "string", "format": "date-time"}, "updatedAt": {"type": "string", "format": "date-time"}}},
    "AudioFormat": {"type": "string", "enum": ["mp3", "flac"]},
    "AdminTrackList": {"type": "object", "required": ["data"], "properties": {"data": {"type": "array", "items": {"$ref": "#/components/schemas/AdminTrack"}}}},
    "AdminTrackResponse": {"type": "object", "required": ["data"], "properties": {"data": {"$ref": "#/components/schemas/AdminTrack"}}},
...
    "Problem": {"type": "object", "required": ["type", "title", "status", "detail"], "properties": {"type": {"type": "string", "format": "uri"}, "title": {"type": "string"}, "status": {"type": "integer"}, "detail": {"type": "string"}}}
  }}
}